+++ src/navigation/AuthTabs.tsx
@@ -1,0 +1,20 @@
+import React from 'react';
+import {createNativeStackNavigator} from '@react-navigation/native-stack';
+import LoginScreen from '../screens/login/login';
+import RegisterScreen from '../screens/register/register';
+
+type AuthStackParamList = {
+  Login: undefined;
+  Register: undefined;
+};
+
+const AuthStack = createNativeStackNavigator<AuthStackParamList>();
+
+const AuthTabs = () => (
+  <AuthStack.Navigator>
+    <AuthStack.Screen name="Login" component={LoginScreen} />
+    <AuthStack.Screen name="Register" component={RegisterScreen} />
+  </AuthStack.Navigator>
+);
+
+export default AuthTabs;
